requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
colorama==0.4.3
config==0.5.0.post0
idna==2.8
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
pylint==2.4.4
python-decouple==3.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
typed-ast==1.4.1
urllib3==1.25.7
wrapt==1.11.2